Are these changes okay?
Chad and I were retesting these protocols for tomorrow, but we realized that we don't want the session to end for 10 disengaged trials quite yet. We also want the protocol to add a time-out if the animal engages at the center but then refuses to make a choice. Is there any way that these changes could break anything else?